app/requirements.txt
flet
absl-py==2.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
ctranslate2==4.7.1
cuda-pathfinder==1.4.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastar==0.8.0
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.24.3
flake8==4.0.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iso8601==2.1.0
isort==8.0.0
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.6.1
mdurl==0.1.2
mediapipe==0.10.9
moviepy==2.2.1
mpmath==1.3.0
mypy_extensions==1.1.0
numpy==2.4.2
nvidia-nvjitlink-cu12==12.8.93
onnxruntime==1.24.2
openai==2.24.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pathspec==1.0.4
pillow==11.3.0
platformdirs==4.9.2
proglog==0.1.12
protobuf==3.20.3
pycodestyle==2.8.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==2.4.0
Pygments==2.19.2
pyparsing==3.3.2
pypika-tortoise==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
starlette==0.52.1
sympy==1.14.0
tokenizers==0.22.2
tomli==2.4.0
tortoise-orm==1.1.5
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yt-dlp==2026.2.21
